[fix] #253 : 모든 CD 테두리 radius 값 추가
